+++ src/Components/Category/CategoryBase.tsx
@@ -2,10 +2,12 @@ import { ButtonBase, createStyles, makeStyles } from '@material-ui/core'
 import { ButtonBaseProps } from '@material-ui/core/ButtonBase'
 import React, { FC } from 'react'
 
+import { BORDER_RADIUS_HUGE } from '../../theme'
+
 const useStyles = makeStyles(() =>
     createStyles({
         buttonBase: {
-            borderRadius: 16,
+            borderRadius: BORDER_RADIUS_HUGE,
         },
     })
 )
