+++ src/Components/Shared/Comments/Comments.tsx
@@ -25,6 +25,7 @@ export const Comments: FC<CommentsDocument & CommentsCollections> = ({
 
             <CommentsDrawer
                 collection={collection}
+                numberOfComments={numberOfComments}
                 name={name}
                 open={drawer}
                 onClose={handleDrawerChange}
